Add Movie model and populate db with 20 movies from The Movie Database API.
Created a movie.js file that exports a Movie mongoose model. Used
that model in populate-db.js to create a Movie from each JSON
object returned from the HTTP call to The Movie Database API and then
stored that Movie in the database. The logic for populating a the
movies in the db is now complete.