Reworked the API Key section of the README.
I think I want to have a 'root' api key that is used to hit the
/users routes and then api key fields on each user model that are
used to hit the /movies routes. I'm thinking that the 'root' api key
is basically like the admin account that can deal with user and movie
CRUD operations, while individual user accounts can only deal with
editing their own user account and editing movies in their section
of the db.