Refactor error-handling logic in POST /users.
I'm not sure if this is the best strategy, but I'm handling the
error-handling in 3 places. For HTTP request-related errors, I'm
returning the JSON response in the route-handler. For Model-related
errors, I'm creating and passing the JSON response back from the
given CRUD route, so no indication is given in the route-handler, the
response object itself will know about the error. For db-level errors,
I'm passing the error along to next() which will run app-level
error-handling middleware. This seems logical at this point, but
also messy as there are different areas where error-response objects
are being created and returned from.