Make all fields of User model required and add error-handling if a validation error occurs.
I want all of the fields of a user to be required in order to create
a new user. Instead of running validations when the model is saved
to the db, I chose to run them separately, so I could return a proper
error object. This also led me to refactor the src/db/db.js file a
little, as well as the route-handler.