Return all movies at GET /movies.
Refactored the way the read route is going to work a little bit.
Intead of sending a json response from the db method, I'm just
returning the array of documents, as there is no real validation
that is occuring in the db file so there is no reason to return
a statusCode. Am not authenticating the API Key yet, so as long
as the apiKey appears in the query string the GET /movies route
will work.