Refactor response from db.read() in GET /movies/:id.
Originally, I was sending back the movie or an empty array from
db.read('movies', :id), but this presented a problem for error-handling.
When an invalid id is passed in, as long as it is a valid ObjectID,
the method worked as intended. If an id was passed in that wasn't
a valid ObjectID, an error was thrown that was picked up by the
app-level error handler. The best way to deal with this seemed to
be to handle validating the id in the db.read() method, and sending
back a response with statusCode and statusMessage properties when
there was an error, or statusCode and movie properties on success.