Implement basic version of GET /users/:id.
At this point, as long as the apiKey passed in with the request
is a valid apiKey, the user will be returned. I need to restrict
this route to only return the user when the apiKey matches the
apiKey that the user with the give :id has, or is the root apiKey.