Refactor users and movies routers to use a call to router.param to handle some :id validation.
This is a big win for DRYness. I had a bunch of duplication in the
db.js file between the read, update, and delete CRUD actions, all
relating to validating whether an id was a valid ObjectID and
whether a resource existed with that given id. I was able to
move this logic from each method into a router.param('id') call
in the users and movies router. This still isn't great, but it
is much cleaner.