Change router.use() to router.get('/') at GET /users.
I was using a pattern which I didn't like in the users router. I
feel like I should only use router.use() when some peice of middleware
will apply to multiple routes. In this instance, the GET /users route
is restricted to the root user, but I was calling a validation function
from router.use(), and just had it at the end of the file, right
above the route handler. Now that it is scoped to the method, I
can move the route around.