Refactor routers to remove redunant error-handling checks.
Now that I'm validating id's in middleware and returning a
response, I don't need to check the dbResponse after calling
db.read or db.delete. Before, I was sending back an error
object that needed to be handled in the route handler by checking
the dbResponse.statusCode property, but this has all been moved
to the router.param('id') call.