Refactor how validatePaths is called in PATCH /users/:id.
I was calling validating that the updated user being passed in
didn't have any invalid field names in the actual db.update method,
but I've moved this validation up to a middleware validator in
validate.js. Going to refactor how the db.create method works
as well, since it uses the same validation.