Refactor how I'm validating field names at Create and Update resource.
Before, I was validating the field names in the db.create and
db.update methods. This was a little repetitive, and it makes
more sense for this to be a middleware validation function anyway
since the response needs to be sent when this validation fails. Now,
the POST /resource and PATCH /resource/:id route-handlers both have
to call the validate.fieldNames middleware, which I don't love, but
there is no way at this point that I can see to have these calls be
placed in one location and still have the same effect.