Refactor db.js and routers to peform model validations as a middleware call.
I was validating that the model coming in to the db.create method
had valid values for fields in the db.create method itself. I've
moved this validation out into validate.js so that I can call
it as middleware. Now, it has to appear in the route-handler
signature at router.post('/') and router.patch('/:id') for all
resources, which seems like a better pattern than having it buried
in the db.create call.