Refactor response from db.read() to be similar to other CRUD calls.
db.read() was the only CRUD method that would return only a resource
instead of an object with a statusCode and resource. I want the
route-handlers to all look and behave the same, so I changed db.read
a bit so that it always returns an object with a statusCode and resource,
and then the response is sent res.status(dbResponse.statusCode).json(dbResponse.resource).
This just looks cleaner, and keeps all the routes similar, which I
like.