Stack calls to validate api keys in router.use() instead of calling router.use() twice.
I was using an old pattern of calling multiple versions of the
same method with different middleware functions instead of just
passing all relevant middleware functions to the route handler. In
this instance, it was the router.use() calls that validate the api
keys in both movies and users routers.