Add commment about router.all() use.
I tried to use router.use() in place of router.all(), and realized
that since I'm using req.baseUrl in the middleware this would cause
an issue. router.use() considers the baseUrl to be the entire url,
while router.all() behaves like router.method(), and baseUrl is
/resource. I need this behaviour, as I'm determining which Model
is being validated based on the value of baseUrl.